[FLINK-19698] Move the CheckpointListener from flink-runtime to flink-core. To maintain backwards compatibility, the CheckpointListener in flink-runtime is not deleted immediately.
