+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/CheckpointListener.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -24,28 +24,13 @@ import org.apache.flink.annotation.Public;
  * This interface must be implemented by functions/operations that want to receive
  * a commit notification once a checkpoint has been completely acknowledged by all
  * participants.
+ *
+ * @deprecated This interface has been moved to
+ * {@link org.apache.flink.api.common.state.CheckpointListener}. This class is kept
+ * to maintain backwards compatibility and will be removed in future releases.
  */
 @Public
-public interface CheckpointListener {
+@Deprecated
+public interface CheckpointListener extends org.apache.flink.api.common.state.CheckpointListener {
 
-	/**
-	 * This method is called as a notification once a distributed checkpoint has been completed.
-	 *
-	 * <p>Note that any exception during this method will not cause the checkpoint to
-	 * fail any more.
-	 *
-	 * @param checkpointId The ID of the checkpoint that has been completed.
-	 * @throws Exception This method can propagate exceptions, which leads to a failure/recovery for
-	 *                   the task. Not that this will NOT lead to the checkpoint being revoked.
-	 */
-	void notifyCheckpointComplete(long checkpointId) throws Exception;
-
-	/**
-	 * This method is called as a notification once a distributed checkpoint has been aborted.
-	 *
-	 * @param checkpointId The ID of the checkpoint that has been aborted.
-	 * @throws Exception This method can propagate exceptions, which leads to a failure/recovery for
-	 *                   the task.
-	 */
-	default void notifyCheckpointAborted(long checkpointId) throws Exception {}
 }
